move downsample_contour_points
