+++ src/types/native.rs
@@ -11,7 +11,7 @@ pub mod record;
 pub use centerline::Centerline;
 pub use centerline_point::CenterlinePoint;
 pub use contour::{Contour, ContourType};
-pub use contour_point::ContourPoint;
+pub use contour_point::{downsample_contour_points, ContourPoint};
 pub use frame::Frame;
 pub use geometry::Geometry;
 pub use geometry_pair::GeometryPair;
